Add handling keys from the Minimal Phone MP01.
This includes left+right shift and left alt, as well as special handling
for the Emoji and Dictate/period keys. The MP01 also lacks a dedicated
control key, so this includes treating the Dictate/period key like
control when it's held before pressing another key.

The "0" alt-key behavior of the emoji key is included, but the main
function does nothing in most apps so far.

This also adds a separate "caps" state tracking from "shift", so the
autocaps doesn't cause the "." key to output ">".

Most of this won't be needed on the Titan, but hasn't been tested.